+++ Automate/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Optimized to avoid rescan when a machine is removed from an inactive machine group.
+
 ## 1.27.0
 Released 05 June 2022 for SMAPI 3.14.0 or later.